requirements.txt
amqp==5.3.1
attrs==24.2.0
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
Flask==3.1.0
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.0
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
protobuf==5.28.3
PyMatting==1.1.13
python-dateutil==2.9.0.post0
redis==5.2.0
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sympy==1.13.3
tifffile==2024.9.20
tqdm==4.67.0
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
